requirements.txt
ultralytics>=8.1.0
opencv-python>=4.6.0
numpy>=1.24.0
matplotlib>=3.7.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.4.0
pyyaml>=6.0
tqdm>=4.65.0
albumentations>=1.3.1 